+++ packages/media-parser/src/boxes/iso-base-media/moov/moov.ts
@@ -4,7 +4,7 @@ import type {Options, ParseMediaFields} from '../../../options';
 import type {AnySegment, IsoBaseMediaBox} from '../../../parse-result';
 import type {ParserState} from '../../../state/parser-state';
 import type {BaseBox} from '../base-type';
-import {parseIsoBaseMediaBoxes} from '../process-box';
+import {parseIsoBaseMediaBoxes} from '../parse-boxes';
 
 export interface MoovBox extends BaseBox {
 	type: 'moov-box';
