+++ packages/media-parser/src/boxes/iso-base-media/trak/trak.ts
@@ -4,7 +4,7 @@ import type {Options, ParseMediaFields} from '../../../options';
 import type {AnySegment, IsoBaseMediaBox} from '../../../parse-result';
 import type {ParserState} from '../../../state/parser-state';
 import type {BaseBox} from '../base-type';
-import {parseIsoBaseMediaBoxes} from '../process-box';
+import {parseIsoBaseMediaBoxes} from '../parse-boxes';
 
 export interface TrakBox extends BaseBox {
 	type: 'trak-box';
